+++ src/Query/SelectQuery.php
@@ -208,6 +208,34 @@ class SelectQuery extends AbstractQuery
     }
 
     /**
+     * Adds an exists condition to the query.
+     *
+     * @param callable $f
+     *
+     * @return self
+     */
+    public function exists(callable $f)
+    {
+        $this->where->addExistsCondition($f);
+
+        return $this;
+    }
+
+    /**
+     * Adds a not exists condition to the query.
+     *
+     * @param callable $f
+     *
+     * @return self
+     */
+    public function notExists(callable $f)
+    {
+        $this->where->addNotExistsCondition($f);
+
+        return $this;
+    }
+
+    /**
      * Sets the limit for the query.
      *
      * @param int $limit