+++ tests/Query/SelectQueryTest.php
@@ -90,6 +90,26 @@ class SelectQueryTest extends PHPUnit_Framework_TestCase
         $this->assertEquals([['date', 'NOT BETWEEN', 2015, 2016]], $query->getWhere()->getConditions());
     }
 
+    public function testExists()
+    {
+        $query = new SelectQuery();
+
+        $f = function (SelectQuery $query) {};
+
+        $this->assertEquals($query, $query->exists($f));
+        $this->assertEquals([['EXISTS', $f]], $query->getWhere()->getConditions());
+    }
+
+    public function testNotExists()
+    {
+        $query = new SelectQuery();
+
+        $f = function (SelectQuery $query) {};
+
+        $this->assertEquals($query, $query->notExists($f));
+        $this->assertEquals([['NOT EXISTS', $f]], $query->getWhere()->getConditions());
+    }
+
     public function testLimit()
     {
         $query = new SelectQuery();