+++ tests/Query/UpdateQueryTest.php
@@ -8,6 +8,7 @@
  * @copyright 2015 Jared King
  * @license MIT
  */
+use JAQB\Query\SelectQuery;
 use JAQB\Query\UpdateQuery;
 
 class UpdateQueryTest extends PHPUnit_Framework_TestCase
@@ -81,6 +82,26 @@ class UpdateQueryTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals($query, $query->notBetween('date', 2015, 2016));
         $this->assertEquals([['date', 'NOT BETWEEN', 2015, 2016]], $query->getWhere()->getConditions());
+    }
+
+    public function testExists()
+    {
+        $query = new UpdateQuery();
+
+        $f = function (SelectQuery $query) {};
+
+        $this->assertEquals($query, $query->exists($f));
+        $this->assertEquals([['EXISTS', $f]], $query->getWhere()->getConditions());
+    }
+
+    public function testNotExists()
+    {
+        $query = new UpdateQuery();
+
+        $f = function (SelectQuery $query) {};
+
+        $this->assertEquals($query, $query->notExists($f));
+        $this->assertEquals([['NOT EXISTS', $f]], $query->getWhere()->getConditions());
     }
 
     public function testOrderBy()