+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Support for BETWEEN and NOT BETWEEN as WHERE conditions.
 - Added NOT where conditions.
 - Added OR where conditions.
+- Added EXISTS where conditions.
 
 ### Changed
 - SetStatement `getValues()` now returns parameterized values and `getSetValues()` returns the key-value map of values being set.